// frontend/api/models/Media.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Media = {
  /**
   * @type string, date-time
   */
  created_at: string;
  /**
   * @type string
   */
  id: string;
  /**
   * @description The media type of the object
   * @type string
   */
  media_type: string;
  /**
   * @description The S3 key of the media object
   * @type string
   */
  s3key: string;
  /**
   * @description The title of the media
   * @type string
   */
  title: string;
};